geo-data-resolver/requirements.txt
flask
requests
PyYAML
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub
google-cloud-logging
google-cloud-secret-manager
gunicorn
flask-talisman
pandas
grpcio==1.74.0
googlemaps